+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/manage/ThrottlingRate.tsx
@@ -14,7 +14,7 @@ import { toast, ToastContainer } from "@formBuilder/components/shared/Toast";
 import { getSecondsInWeeks, getWeeksInSeconds } from "@lib/utils/date/dateConversions";
 import { SubmitButton } from "@clientComponents/globals/Buttons/SubmitButton";
 import { Tooltip } from "@formBuilder/components/shared/Tooltip";
-import { useRehydrate } from "@lib/store/useTemplateStore";
+import { useRehydrate } from "@lib/store/hooks/useRehydrate";
 import Skeleton from "react-loading-skeleton";
 import { formClosingDateEst } from "@lib/utils/date/utcToEst";
 import { logMessage } from "@lib/logger";
