+++ lib/hooks/form-builder/useTemplateContext.tsx
@@ -3,7 +3,8 @@ import React, { createContext, useState, useContext, useRef } from "react";
 import { logMessage } from "@lib/logger";
 import { CreateOrUpdateTemplateType, createOrUpdateTemplate } from "@formBuilder/actions";
 import { FormRecord } from "@lib/types";
-import { useSubscibeToTemplateStore, useTemplateStore } from "@lib/store/useTemplateStore";
+import { useTemplateStore } from "@lib/store/useTemplateStore";
+import { useSubscibeToTemplateStore } from "@lib/store/hooks/useSubscibeToTemplateStore";
 
 interface TemplateApiType {
   templateIsDirty: React.MutableRefObject<boolean>;
